fix bug

+++ router/ant/src/lib.rs
@@ -483,13 +483,14 @@ impl AntRouterInner {
         end: AccountAddress,
     ) -> Result<Vec<BalanceQueryResponse>> {
         let mut balance_map = HashMap::new();
-        let mut min_pressure = std::u64::MAX;
+        let mut min_pressure = std::i128::MAX;
         for path in paths.into_iter() {
             let balances = self.format_response_list(path, start, end);
-            let mut pressure: u64 = 0;
+            let mut pressure: i128 = 0;
             for balance in balances.iter() {
-                pressure = pressure + balance.total_pay_amount + balance.remote_balance
-                    - balance.local_balance;
+                pressure =
+                    pressure + balance.total_pay_amount as i128 + balance.remote_balance as i128
+                        - balance.local_balance as i128;
             }
             balance_map.insert(pressure, balances);
             if pressure < min_pressure {
